check

+++ Vol.1/src/server.c
@@ -26,6 +26,7 @@ sigset_t setusr1;
 struct sigaction usr2;
 sigset_t setusr2;
 
+int client_id = -1;
 int msqid = -1;
 int games = GAMES;
 int server_socket_read = -1;
@@ -53,6 +54,8 @@ void sigchld_handler(int s, siginfo_t *info, void *param) {
   pid_t pid = info->si_pid;
   int sid = -1;
   int id = -1;
+  int w = -1;
+  int l = -1;
   ipc_t *p = NULL;
   games_t *g = NULL;
   pthread_mutex_lock(&mut);
@@ -78,34 +81,42 @@ void sigchld_handler(int s, siginfo_t *info, void *param) {
             // Win player2
             sbuf.mtype = WIN_END_GAME;
             sprintf(sbuf.mtext, "win");
-            msgsnd(g->player2->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
-            fprintf(stderr,
-                    "Send msg [fd: %d | msg: %s]\n",
-                    g->player2->msqid,
-                    sbuf.mtext);
-            kill(g->player2->pid, SIGUSR1);
+            if (g->player2 != NULL) {
+              msgsnd(g->player2->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+              /*  fprintf(stderr, "Send msg [fd: %d | msg: %s]\n",
+                 g->player2->msqid, sbuf.mtext);*/
+              id = g->id;
+              w = g->player2_id;
+              l = g->player1_id;
+              kill(g->player2->pid, SIGUSR1);
+            }
             remove_game(
                     &game_ch,
                     g->id,
-                    g->player2->id,
-                    g->player1->id,
+                    g->player2_id,
+                    g->player1_id,
                     &games_mutex);
-          } else {
+            synchronized_game_w(id, w, l);
+          } else if (pid == g->player2->pid) {
             // Win player1
             sbuf.mtype = WIN_END_GAME;
             sprintf(sbuf.mtext, "win");
-            msgsnd(g->player1->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
-            fprintf(stderr,
-                    "Send msg [fd: %d | msg: %s]\n",
-                    g->player1->msqid,
-                    sbuf.mtext);
-            kill(g->player1->pid, SIGUSR1);
+            if (g->player1 != NULL) {
+              msgsnd(g->player1->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+              /*    fprintf(stderr, "Send msg [fd: %d | msg: %s]\n",
+                 g->player1->msqid, sbuf.mtext);*/
+              id = g->id;
+              w = g->player1_id;
+              l = g->player2_id;
+              kill(g->player1->pid, SIGUSR1);
+            }
             remove_game(
                     &game_ch,
                     g->id,
-                    g->player1->id,
-                    g->player2->id,
+                    g->player1_id,
+                    g->player2_id,
                     &games_mutex);
+            synchronized_game_w(id, w, l);
           }
         }
       }
@@ -119,7 +130,11 @@ void sigusr1_handler(int s, siginfo_t *info, void *param) {
   pid_t pid = info->si_pid;
   int id = -1;
   int fd = -1;
+  int fi = -1;
+  int si = -1;
+  int own = -1;
   ipc_t *p = NULL;
+  games_t *g = NULL;
   ipc_t *ptr = NULL;
   char buffer[MAXDATASIZE];
   fprintf(stderr,
@@ -129,15 +144,12 @@ void sigusr1_handler(int s, siginfo_t *info, void *param) {
           mainpid);
   if (pid == mainpid) {
     msgrcv(msqid, &rbuf, sizeof(message_t), 0, 0);
-    fprintf(stderr,
-            "Receive msg [fd: %d | msg: %s | type: %ld]\n",
-            msqid,
-            rbuf.mtext,
-            rbuf.mtype);
+    /*    fprintf(stderr, "Receive msg [fd: %d | msg: %s | type: %ld]\n", msqid,
+                rbuf.mtext, rbuf.mtype);*/
     if (rbuf.mtype == WIN_END_GAME) {
       fprintf(stderr, "Server send win [id: %d]\n", msqid);
       send_packet(make_packet(SERVICE, id, 0, "winner"));
-    } else if (rbuf.mtype == SINCHRONIZE_GAME) {
+    } else if (rbuf.mtype == SYNCHRONIZE_GAME) {
       fprintf(stderr,
               "Request for sinchronisation [pid: %d | id: %d | state: %d]\n",
               getpid(),
@@ -145,13 +157,20 @@ void sigusr1_handler(int s, siginfo_t *info, void *param) {
               check_connection());
       if (check_connection()) {
         sprintf(buffer, "%d.game", atoi(rbuf.mtext));
-        fd = open(buffer, O_RDONLY);
-        read(fd, buffer, sizeof(int) * 5 + strlen("AAAAAAAAA") + 1);
-        close(fd);
-        send_packet(make_packet(
-                SINCHRONIZE, this_server->number, atoi(rbuf.mtext), buffer));
+        if (!access(buffer, 0)) {
+          fd = open(buffer, O_RDONLY);
+          read(fd, buffer, SIZE_GAME);
+          close(fd);
+          send_packet(make_packet(
+                  SYNCHRONIZE, this_server->number, atoi(rbuf.mtext), buffer));
+        } else {
+          sscanf(rbuf.mtext, "%d %d %d", &id, &fi, &si);
+          sprintf(buffer, "%d %d %d", id, fi, si);
+          send_packet(make_packet(
+                  SYNCHRONIZE_REMOVE, this_server->number, id, buffer));
+        }
         fprintf(stderr,
-                "Sinchronize! [pid: %d | id: %d]\n",
+                "Synchronize! [pid: %d | id: %d]\n",
                 getpid(),
                 atoi(rbuf.mtext));
       } else {
@@ -160,6 +179,13 @@ void sigusr1_handler(int s, siginfo_t *info, void *param) {
                 getpid(),
                 atoi(rbuf.mtext));
       }
+    } else if (rbuf.mtype == START_GAME) {
+      fprintf(stderr,
+              "Connect to game [pid: %d | id: %d]\n",
+              getpid(),
+              atoi(rbuf.mtext));
+      send_packet(
+              make_packet(SERVICE, client_id, atoi(rbuf.mtext), "start_game"));
     }
   } else {
     p = get_msq_pid(&msq, pid, FALSE, &msq_mutex);
@@ -169,6 +195,7 @@ void sigusr1_handler(int s, siginfo_t *info, void *param) {
             p->msqid,
             rbuf.mtext,
             rbuf.mtype);
+
     if (rbuf.mtype == NONE_CMD) {
       if (!strcmp(rbuf.mtext, "get_id")) {
         while (TRUE) {
@@ -189,17 +216,57 @@ void sigusr1_handler(int s, siginfo_t *info, void *param) {
         sbuf.mtype = SET_ID;
         sprintf(sbuf.mtext, "%d", p->id);
         msgsnd(p->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
-        fprintf(stderr, "Send msg [fd: %d | msg: %s]\n", p->msqid, sbuf.mtext);
+        /*  fprintf(stderr, "Send msg [fd: %d | msg: %s]\n", p->msqid,
+         * sbuf.mtext);*/
         fprintf(stderr,
                 "Clients [count: %ld(%ld)]\n",
                 clients_available,
                 clients);
       }
+    } else if (rbuf.mtype == RECONNECT_ID) {
+      id = atoi(rbuf.mtext);
+      fprintf(stderr, "Reconnect id [id: %d]\n", id);
+      if (p != NULL)
+        p->id = id;
+      pthread_mutex_lock(&clients_av_mutex);
+      ++clients_available;
+      ++clients;
+      pthread_mutex_unlock(&clients_av_mutex);
+      print_msq(&msq, &msq_mutex);
     } else if (rbuf.mtype == SERVER_SET_ID) {
       id = atoi(rbuf.mtext);
       p = get_msq_pid(&msq, pid, FALSE, &msq_mutex);
       p->id = id;
       fprintf(stderr, "Server set id [id: %d]\n", p->id);
+    } else if (rbuf.mtype == SYNCHRONIZE_CHANGE) {
+      sscanf(rbuf.mtext, "%d %d %d %d", &id, &fi, &si, &own);
+      //      id = atoi(rbuf.mtext);
+      fprintf(stderr, "Synchronize... [id: %d]\n", id);
+      g = get_game_id(&game_ch, id, &games_mutex);
+      if (g == NULL) {
+        add_game(&game_ch, id, NULL, NULL, fi, si, own, &games_mutex);
+      }
+    } else if (rbuf.mtype == SYNCHRONIZE_RM) {
+      sscanf(rbuf.mtext, "%d %d %d", &id, &fi, &si);
+      fprintf(stderr, "Synchronize... [id: %d]\n", id);
+      remove_game(&game_ch, id, fi, si, &games_mutex);
+    } else if (rbuf.mtype == RECONNECT_GAME) {
+      sscanf(rbuf.mtext, "%d %d", &fi, &id);
+      fprintf(stderr, "Connect client to game [game: %d | id: %d]\n", id, fi);
+      g = get_game_id(&game_ch, id, &games_mutex);
+      if (g != NULL) {
+        if (g->player1_id == fi) {
+          p = get_msq_id(&msq, fi, FALSE, &msq_mutex);
+          msq_set_game(&p, g->id, &msq_mutex);
+          g->player1 = p;
+        } else if (g->player2_id == fi) {
+          p = get_msq_id(&msq, fi, FALSE, &msq_mutex);
+          msq_set_game(&p, g->id, &msq_mutex);
+          g->player2 = p;
+        }
+      } else {
+        msq_set_game(&p, -1, &msq_mutex);
+      }
     }
     print_msq(&msq, &msq_mutex);
   }
@@ -343,10 +410,27 @@ void synchronized_game(int game_id) {
   int i = 0;
   for (i = 0; i < known_servers->count; ++i) {
     p = get_msq_id(&msq, cursor->number, TRUE, &msq_mutex);
-    sbuf.mtype = SINCHRONIZE_GAME;
+    sbuf.mtype = SYNCHRONIZE_GAME;
     sprintf(sbuf.mtext, "%d", game_id);
     msgsnd(p->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
-    fprintf(stderr, "Send msg [fd: %d | msg: %s]\n", p->msqid, sbuf.mtext);
+    /*  fprintf(stderr, "Send msg [fd: %d | msg: %s]\n", p->msqid,
+     * sbuf.mtext);*/
+    kill(p->pid, SIGUSR1);
+    cursor = cursor->next;
+  }
+}
+/*---------------------------------------------------------------------------*/
+void synchronized_game_w(int game_id, int w, int l) {
+  srv_t *cursor = known_servers->srvs;
+  ipc_t *p;
+  int i = 0;
+  for (i = 0; i < known_servers->count; ++i) {
+    p = get_msq_id(&msq, cursor->number, TRUE, &msq_mutex);
+    sbuf.mtype = SYNCHRONIZE_GAME;
+    sprintf(sbuf.mtext, "%d %d %d", game_id, w, l);
+    msgsnd(p->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+    /*    fprintf(stderr, "Send msg [fd: %d | msg: %s]\n", p->msqid,
+     sbuf.mtext);*/
     kill(p->pid, SIGUSR1);
     cursor = cursor->next;
   }
@@ -357,11 +441,28 @@ void *manager() {
   ipc_t *f = NULL;
   ipc_t *s = NULL;
   while (!manager_join) {
-    if (clients_available > 0 && !(clients_available % 2) &&
-        games_curr <= GAMES) {
+    sleep(1);
+    if (clients_available > 1 && games_curr <= GAMES) {
       game_id = rand() % 100;
       get_free_pair_msq(&msq, &f, &s, &msq_mutex);
-      add_game(&game_ch, game_id, f, s, this_server->number, &games_mutex);
+      add_game(
+              &game_ch,
+              game_id,
+              f,
+              s,
+              f->id,
+              s->id,
+              this_server->number,
+              &games_mutex);
+
+      sbuf.mtype = START_GAME;
+      sprintf(sbuf.mtext, "%d", game_id);
+      msgsnd(f->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+      kill(f->pid, SIGUSR1);
+      sbuf.mtype = START_GAME;
+      sprintf(sbuf.mtext, "%d", game_id);
+      msgsnd(s->msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+      kill(s->pid, SIGUSR1);
       pthread_mutex_lock(&clients_av_mutex);
       clients_available -= 2;
       pthread_mutex_unlock(&clients_av_mutex);
@@ -374,14 +475,15 @@ void *manager() {
 /*---------------------------------------------------------------------------*/
 void client_connection() {
   pid_t pid;
-  int client_id = -1;
+  int fd = -1;
+  int game_id = -1;
   pthread_t reader_tid = -1;
   pthread_t writer_tid = -1;
   pthread_attr_t reader_attr;
   pthread_attr_t writer_attr;
   key_t key;
   char str[MAXDATASIZE];
-
+  int t[5] = {0};
   state_connection = CONN_TRUE;
   pid = getpid();
   key = pid;
@@ -416,12 +518,20 @@ void client_connection() {
       } else if (p.type == CONN_CLIENT) {
         if (p.client_id > 0) {
           client_id = p.client_id;
+          if (p.packet_id > 0) {
+            game_id = p.packet_id;
+            sbuf.mtype = RECONNECT_GAME;
+            sprintf(sbuf.mtext, "%d %d", client_id, game_id);
+            msgsnd(msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+            kill(mainpid, SIGUSR1);
+          }
         }
         if (client_id == -1) {
           sbuf.mtype = NONE_CMD;
           strcpy(sbuf.mtext, "get_id");
           msgsnd(msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
-          fprintf(stderr, "Send msg [fd: %d | msg: %s]\n", msqid, sbuf.mtext);
+          //          fprintf(stderr, "Send msg [fd: %d | msg: %s]\n", msqid,
+          //          sbuf.mtext);
           kill(mainpid, SIGUSR1);
           msgrcv(msqid, &rbuf, MAXDATASIZE, SET_ID, 0);
           fprintf(stderr,
@@ -430,6 +540,12 @@ void client_connection() {
                   rbuf.mtext,
                   rbuf.mtype);
           client_id = atoi(rbuf.mtext);
+        } else {
+          sbuf.mtype = RECONNECT_ID;
+          sprintf(sbuf.mtext, "%d", client_id);
+          msgsnd(msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+          fprintf(stderr, "Send msg [fd: %d | msg: %s]\n", msqid, sbuf.mtext);
+          kill(mainpid, SIGUSR1);
         }
         send_packet(make_packet(SERVICE, client_id, rand() % 1000, "set_id"));
         fprintf(stderr,
@@ -455,12 +571,33 @@ void client_connection() {
                         "Receive SERVICE packet [id: %d | msg: %s]\n",
                         p.client_id, p.buffer);
         */
-      } else if (p.type == SINCHRONIZE) {
+      } else if (p.type == SYNCHRONIZE) {
         fprintf(stderr,
-                "Receive SINCHRONIZE packet [id: %d | game: %d | msg: %s]\n",
+                "Receive SYNCHRONIZE packet [id: %d | game: %d | msg: %s]\n",
                 p.client_id,
                 p.packet_id,
                 p.buffer);
+        sprintf(str, "%d.game", p.packet_id);
+        fd = open(str, O_WRONLY | O_CREAT, 0666);
+        write(fd, p.buffer, SIZE_GAME);
+        close(fd);
+        sbuf.mtype = SYNCHRONIZE_CHANGE;
+        memcpy(t, p.buffer, 5 * sizeof(int));
+        sprintf(sbuf.mtext, "%d %d %d %d", p.packet_id, t[2], t[3], t[1]);
+        msgsnd(msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+        kill(mainpid, SIGUSR1);
+      } else if (p.type == SYNCHRONIZE_REMOVE) {
+        fprintf(stderr,
+                "Receive SYNCHRONIZE_REMOVE packet [id: %d | game: %d | msg: "
+                "%s]\n",
+                p.client_id,
+                p.packet_id,
+                p.buffer);
+        sbuf.mtype = SYNCHRONIZE_RM;
+        sscanf(p.buffer, "%d %d %d", &t[1], &t[2], &t[3]);
+        sprintf(sbuf.mtext, "%d %d %d", p.packet_id, t[2], t[3]);
+        msgsnd(msqid, &sbuf, MAXDATASIZE, IPC_NOWAIT);
+        kill(mainpid, SIGUSR1);
       }
     }
   }
